[REG-1291] Overlay support for Unity local bots (#47)

+++ Runtime/Scripts/RGBotConfigs/RGAgent.cs
@@ -15,7 +15,7 @@ namespace RegressionGames.RGBotConfigs
 
         // The client Id that owns this agent
         // Used as a performance optimization for mapping the ClientId into the state payloads
-        public uint? ClientId = null;
+        public long? ClientId = null;
 
         public RGAgent()
         {
